docs: Move documentation comments into header files
In my testing, doxygen did not show the comments if they were in `.tpp` files.
VS Code with clangd seems to use the comment of the function definition though, so it's a trade-off.

My approach now will be to put "public" documentation in the header, and implementation details in the source files.