build: Include dependency files always unless known they are not needed
Changing from a whitelist to a blacklist avoids the issue that the dependency tracking doesn't work for build targets a user of `build.mk` added to their own Makefile.

This change means that dependency files will get built even for unknown targets, but optimizing for misuse is not the goal.

The `MODE` variable now also gets checked.