chore: Add IWYU private pragmas to `.tpp` files
This should signal that the `.tpp` files should not be included directly, but through their corresponding `.hpp` headers.
`clangd` and `clang-tidy` don't seem to issue any warnings atm, but that seems to be a bug. Once that's fixed the IWYU pragma should be a good enough solution to prevent accidental private includes.

`#error` was considered, but `clangd` would always mark it as an error because from the perspective of the file that has the `#error` it is an error. I don't think there is a way to tell `clangd` to always analyze a file as if it was included by another one.