devops: Reenable `clang-tidy` warning for implicit ptr-bool conversions
Implicit conversion from pointers to `bool` was only desired for `Optional` and `Expected`, but the cost of no warnings for normal pointers is not worth it when using `has_value()` is clearer as well.
Users of the library can configure this differently in their own config if implicit `bool` checks on `Optional` and `Expected` are preferred.