fix(`type_traits`): Fix mismatch of boolean logic and name
I tried to make `is_arrayable` a public type trait, but it would have
needed pretty much the same code as `is_abstract`. Most of the code
could be moved to `is_arrayable`, but the existing documentation of
previous research really applies to `is_abstract` and I deem it not
worthwhile to invest much effort into restructuring and retesting for a
non-standard type trait.