refactor: Use common swap pattern instead of `ft::swap`
This is for consistency - switching between libftpp and stdlib containers and algorithms should not change the behavior unexpectedly.
It is also better in the long term to not rely on `ft::swap` because for the common pattern to work one needs to make a non-member swap function for custom types, not member functions.
`ft:swap` can be a useful utility for last resorts when a type only has a member `swap` function and no non-member `swap` can be safely added.