refactor(`algorithm`)!: Rename `swap` to `member_swap`
The name `swap` causes ambiguities when the common `using std::swap;` pattern is used with libftpp types that don't themselves have a global `swap` overload.
ADL finds the fully generic `swap` in the `ft` namespace and it cannot be differentiated from the also fully generic `swap` in the `std` namespace anymore.

A benefit of this renaming is that in case there is no member function `swap`, the common `using std::swap;` pattern can now be used as a fallback.